Move config structs to separate files/their packages
