+++ internal/config/config.go
@@ -10,30 +10,37 @@ import (
 	"github.com/chubin/wttr.in/internal/ip"
 	"github.com/chubin/wttr.in/internal/location"
 	"github.com/chubin/wttr.in/internal/logging"
-	"github.com/chubin/wttr.in/internal/renderer/subprocess"
+	"github.com/chubin/wttr.in/internal/renderer"
 	"github.com/chubin/wttr.in/internal/server"
 	"github.com/chubin/wttr.in/internal/uplink"
 	"github.com/chubin/wttr.in/internal/weather"
 )
 
+// Config is the root configuration structure for the entire service.
 type Config struct {
-	Geo     *location.Config
-	IP      *ip.Config
-	Weather struct {
-		WWO *weather.WWOConfig
-	}
-	Cache   cache.Config
-	Logging logging.Config
-	Uplink  uplink.Config
-	Server  server.Config
+	// Geo contains location-related settings (IP geolocation, default city, etc.)
+	Geo *location.Config `yaml:"geo"`
 
-	// Renderer configuration (currently only subprocess is configurable)
-	Renderer RendererConfig `yaml:"renderer"`
-}
+	// IP contains settings for IP address parsing and geolocation lookup
+	IP *ip.Config `yaml:"ip"`
 
-// RendererConfig holds configuration for all renderers that need external config.
-type RendererConfig struct {
-	Subprocess []subprocess.SubprocessRoute `yaml:"subprocess"`
+	// Weather contains weather data source configuration (WWO, etc.)
+	Weather *weather.Config `yaml:"weather"`
+
+	// Cache defines caching behavior for weather data, location results, etc.
+	Cache *cache.Config `yaml:"cache"`
+
+	// Logging controls log level, format, output destinations, and tracing
+	Logging *logging.Config `yaml:"logging"`
+
+	// Uplink contains configuration for upstream modules
+	Uplink *uplink.Config `yaml:"uplink"`
+
+	// Server defines HTTP server settings (port, timeouts, TLS, etc.)
+	Server *server.Config `yaml:"server"`
+
+	// Renderer configures how responses are rendered (subprocess renderers, templates, etc.)
+	Renderer *renderer.Config `yaml:"renderer"`
 }
 
 // LoadFromYAML loads configuration from a YAML file and returns a pointer to Config
