+++ internal/ip/ip.go
@@ -43,13 +43,6 @@ func (l *Address) String() string {
 		l.CountryCode, l.Country, l.Region, l.City, l.Latitude, l.Longitude)
 }
 
-type Config struct {
-	IPCacheDB   string `yaml:"ipCacheDb"`
-	IPCache     string `yaml:"ipCache"`
-	IPCacheType string `yaml:"ipCacheType"`
-	GeoIP2      string `yaml:"geoip2"`
-}
-
 // Cache provides access to the IP Geodata cache.
 type Cache struct {
 	config *Config
