+++ internal/location/cache.go
@@ -36,18 +36,6 @@ var (
 	defaultFlushInterval = 5000 * time.Millisecond
 )
 
-type Config struct {
-	LocationCacheType string `yaml:"locationCacheType"`
-	LocationCacheDB   string `yaml:"locationCacheDb"`
-	LocationCache     string `yaml:"locationCache"`
-	NominatimServers  []struct {
-		Name  string `yaml:"name"`
-		Type  string `yaml:"type"`
-		URL   string `yaml:"url"`
-		Token string `yaml:"token"`
-	} `yaml:"nominatim"`
-}
-
 // Location represents a cached geocoding result.
 type Location struct {
 	Name     string `db:"name, key" json:"name"`
