+++ internal/weather/config.go
@@ -1,0 +1,5 @@
+package weather
+
+type Config struct {
+	WWO *WWOConfig
+}
